Create initial schedule & agenda (#41)
* feat(web): add rooms

* feat(web): add rooms to sessions

* feat(web): create opening and closing sessions

* feat(web): add hidden sponsors to sessions

As the're commented out they are not visible on the web, let's keep it there for tracking purposes only so we can verify we have everything in place.

* feat(web): create agenda time slots

* feat(web): add scheduled times to sessions

* feat(web): create venue map

Fixes #4

* fix(web): fix session links to rooms

Template always produces lowercase pages.
So in order to linking from map to work properly and still have sessions attached this was necessary.

* fix(web): remove fill from map shapes

* feat(web): update agenda slots to latest plan

* feat(web): update C1 sessions times

* feat(web): update D8 sessions times

* feat(web): update D9 sessions times

* last speakers to agenda

* last speakers to agenda

* last speakers to agenda

---------

Co-authored-by: Martin Damovsky <martin.damovsky@ataccama.com>